# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateHospitalRequest:
    properties:
      address:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  dto.CreatePatientRequest:
    properties:
      dateOfBirth:
        description: 'expected format: YYYY-MM-DD'
        type: string
      email:
        type: string
      firstNameEN:
        type: string
      firstNameTH:
        type: string
      gender:
        enum:
        - M
        - F
        type: string
      lastNameEN:
        type: string
      lastNameTH:
        type: string
      middleNameEN:
        type: string
      middleNameTH:
        type: string
      nationalId:
        type: string
      passportId:
        type: string
      phoneNumber:
        type: string
    required:
    - dateOfBirth
    - firstNameEN
    - firstNameTH
    - gender
    - lastNameEN
    - lastNameTH
    - phoneNumber
    type: object
  dto.CreateStaffRequest:
    properties:
      confirmPassword:
        type: string
      hospitalId:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - confirmPassword
    - hospitalId
    - password
    - username
    type: object
  dto.ErrorResponse:
    properties:
      error:
        example: error message here
        type: string
    type: object
  dto.HospitalResponse:
    properties:
      address:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: string
    type: object
  dto.LoginStaffRequest:
    properties:
      hospitalId:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - hospitalId
    - password
    - username
    type: object
  dto.PaginatedPatientsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.PatientResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  dto.PatientResponse:
    properties:
      createdAt:
        type: string
      dateOfBirth:
        type: string
      email:
        type: string
      firstNameEN:
        type: string
      firstNameTH:
        type: string
      gender:
        type: string
      hospitalId:
        type: string
      id:
        type: string
      lastNameEN:
        type: string
      lastNameTH:
        type: string
      middleNameEN:
        type: string
      middleNameTH:
        type: string
      nationalId:
        type: string
      passportId:
        type: string
      patientHN:
        type: string
      phoneNumber:
        type: string
    type: object
  dto.SearchPatientRequest:
    properties:
      dateOfBirth:
        description: YYYY-MM-DD
        type: string
      email:
        type: string
      firstNameEN:
        type: string
      firstNameTH:
        type: string
      gender:
        type: string
      lastNameEN:
        type: string
      lastNameTH:
        type: string
      limit:
        type: integer
      middleNameEN:
        type: string
      middleNameTH:
        type: string
      nationalId:
        type: string
      page:
        type: integer
      passportId:
        type: string
      patientHN:
        type: string
      phoneNumber:
        type: string
    type: object
  dto.StaffLoginResponse:
    properties:
      token:
        type: string
    type: object
  dto.StaffResponse:
    properties:
      createdAt:
        type: string
      hospitalId:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  dto.UpdateHospitalRequest:
    properties:
      address:
        type: string
      name:
        type: string
    required:
    - name
    type: object
  dto.UpdateStaffRequest:
    properties:
      password:
        description: Optional - only update if provided
        type: string
      username:
        type: string
    required:
    - username
    type: object
host: localhost
info:
  contact:
    email: support@hospital.com
    name: API Support
  description: API for Hospital Middleware System - Patient Information Search
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Hospital Middleware API
  version: "1.0"
paths:
  /hospital:
    get:
      consumes:
      - application/json
      description: Retrieve all hospitals in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.HospitalResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all hospitals
      tags:
      - hospitals
    post:
      consumes:
      - application/json
      description: Register a new hospital in the system
      parameters:
      - description: Hospital details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateHospitalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.HospitalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new hospital
      tags:
      - hospitals
  /hospital/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a hospital by ID
      parameters:
      - description: Hospital ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a hospital
      tags:
      - hospitals
    put:
      consumes:
      - application/json
      description: Update hospital information by ID
      parameters:
      - description: Hospital ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated hospital details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateHospitalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HospitalResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a hospital
      tags:
      - hospitals
  /patients:
    get:
      consumes:
      - application/json
      description: Retrieve patients only within the authenticated staff's hospital
        (no cross-hospital access)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PatientResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get patients by hospital
      tags:
      - patients
    post:
      consumes:
      - application/json
      description: Register a new patient only into the authenticated staff's hospital
        (no cross-hospital access)
      parameters:
      - description: Patient details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePatientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PatientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new patient
      tags:
      - patients
  /patients/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve patient by identifier only within the authenticated staff's
        hospital (no cross-hospital access)
      parameters:
      - description: Patient identifier (nationalId or passportId)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PatientResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search patient by identifier
      tags:
      - patients
  /patients/search:
    post:
      consumes:
      - application/json
      description: Search patients only within the authenticated staff's hospital
        (no cross-hospital access)
      parameters:
      - description: Search filters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SearchPatientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PaginatedPatientsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Search patients with filters
      tags:
      - patients
  /staff:
    get:
      consumes:
      - application/json
      description: Retrieve all staff members
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.StaffResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all staffs
      tags:
      - staffs
  /staff/{id}:
    delete:
      consumes:
      - application/json
      description: Delete staff member by ID
      parameters:
      - description: Staff ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a staff
      tags:
      - staffs
    put:
      consumes:
      - application/json
      description: Update staff information by ID
      parameters:
      - description: Staff ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Updated staff details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StaffResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update a staff member
      tags:
      - staffs
  /staff/create:
    post:
      consumes:
      - application/json
      description: Create a new staff member with credentials
      parameters:
      - description: Staff creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStaffRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.StaffResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new staff
      tags:
      - staffs
  /staff/login:
    post:
      consumes:
      - application/json
      description: Authenticate staff and receive JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginStaffRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StaffLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Staff login
      tags:
      - staffs
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"